fix: #20 출석 체크 로직 이관
