--- src/routes/attendance.api.ts
@@ -1,9 +1,0 @@
-import { Hono } from "hono";
-import attendanceController from "../controllers/attendance.controller";
-import { authMiddleware } from "../middlewares/auth.middleware";
-
-const router = new Hono();
-
-router.post("/", authMiddleware, attendanceController.checkAttendance);
-
-export default router;
