--- src/types/attendance.types.ts
@@ -1,11 +1,0 @@
-import { Context } from "hono";
-
-export interface AttendanceController {
-  checkAttendance: (c: Context) => Promise<Response>;
-}
-
-export interface AttendanceRecordPayload {
-  studentId: string;
-  status: "present" | "absent" | "late";
-  reason?: string;
-}
